pyproject.toml
[project]
name = "simple-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fastembed>=0.7.4",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "livekit-agents[google,silero]>=1.2.8",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "twilio>=9.10.2",
    "uvicorn>=0.41.0",
]
